Module structure proposal
